requirements.txt
bcrypt
Pillow
flet==0.28.3
tensorflow
h5py
numpy
pillow